Serialization for the new item attributes
This commit also cleans up the serialization code in other classes.

+++ src/depotlocker.h
@@ -41,10 +41,10 @@ class DepotLocker final : public Container
 		//serialization
 		Attr_ReadValue readAttr(AttrTypes_t attr, PropStream& propStream) final;
 
-		uint32_t getDepotId() const {
+		uint16_t getDepotId() const {
 			return depotId;
 		}
-		void setDepotId(uint32_t id) {
+		void setDepotId(uint16_t id) {
 			depotId = id;
 		}
 
@@ -60,7 +60,7 @@ class DepotLocker final : public Container
 		}
 
 	private:
-		uint32_t depotId;
+		uint16_t depotId;
 };
 
 #endif
